refactor(storage): change storage folder structure
