+++ modules/storage/src/interfaces/IStorageProvider.ts
