+++ modules/storage/src/interfaces/StorageConfig.ts
