+++ modules/storage/src/interfaces/index.ts
